requirements.txt
numpy
pandas
biopython
matplotlib
future
pyside2
bokeh
